pyproject.toml
[tool.poetry]
name = "chatkdb"
version = "0.1.0"
description = ""
authors = ["Nakaya <eniehack@outlook.jp>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.201"
chromadb = "^0.3.26"
openai = "^0.27.8"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.2"
tiktoken = "^0.4.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
